+++ mage_ai/api/presenters/PipelinePresenter.py
@@ -1,3 +1,4 @@
+from mage_ai.api.operations import constants
 from mage_ai.api.presenters.BasePresenter import BasePresenter
 from mage_ai.data_preparation.models.constants import DATAFRAME_SAMPLE_COUNT_PREVIEW
 
@@ -14,15 +15,36 @@ class PipelinePresenter(BasePresenter):
 
     async def present(self, **kwargs):
         query = kwargs.get('query', {})
-        include_content = query.get('includes_content', [True])
-        if include_content:
-            include_content = include_content[0]
-        include_outputs = query.get('includes_outputs', [True])
-        if include_outputs:
-            include_outputs = include_outputs[0]
 
-        return await self.model.to_dict_async(
-            include_content=include_content,
-            include_outputs=include_outputs,
-            sample_count=DATAFRAME_SAMPLE_COUNT_PREVIEW,
-        )
+        if constants.DETAIL == kwargs['format']:
+            include_content = query.get('includes_content', [True])
+            if include_content:
+                include_content = include_content[0]
+            include_outputs = query.get('includes_outputs', [True])
+            if include_outputs:
+                include_outputs = include_outputs[0]
+
+            return await self.model.to_dict_async(
+                include_content=include_content,
+                include_outputs=include_outputs,
+                sample_count=DATAFRAME_SAMPLE_COUNT_PREVIEW,
+            )
+
+        data = self.model.to_dict()
+
+        include_schedules = query.get('include_schedules', [False])
+        if include_schedules:
+            include_schedules = include_schedules[0]
+
+        if include_schedules:
+            data['schedules'] = self.model.schedules
+
+        return data
+
+
+PipelinePresenter.register_format(
+    constants.LIST,
+    PipelinePresenter.default_attributes + [
+        'schedules',
+    ],
+)
