+++ mage_ai/api/resources/PipelineResource.py
@@ -1,12 +1,89 @@
 from mage_ai.api.resources.GenericResource import GenericResource
-from mage_ai.orchestration.db.models import PipelineSchedule
-from mage_ai.orchestration.db import safe_db_query
 from mage_ai.data_preparation.models.pipeline import Pipeline
+from mage_ai.data_preparation.repo_manager import get_repo_path
+from mage_ai.orchestration.db import safe_db_query
+from mage_ai.orchestration.db.models import PipelineSchedule
 from mage_ai.server.active_kernel import switch_active_kernel
 from mage_ai.server.kernels import PIPELINE_TO_KERNEL_NAME
+from mage_ai.shared.hash import group_by
+from sqlalchemy.orm import aliased
+import asyncio
 
 
 class PipelineResource(GenericResource):
+    @classmethod
+    async def collection(self, query, meta, user, **kwargs):
+        include_schedules = query.get('include_schedules', [False])
+        if include_schedules:
+            include_schedules = include_schedules[0]
+
+        pipeline_uuids = Pipeline.get_all_pipelines(get_repo_path())
+
+        async def get_pipeline(uuid):
+            try:
+                return await Pipeline.get_async(uuid)
+            except Exception:
+                return None
+
+        pipelines = await asyncio.gather(
+            *[get_pipeline(uuid) for uuid in pipeline_uuids]
+        )
+        pipelines = [p for p in pipelines if p is not None]
+
+        @safe_db_query
+        def query_pipeline_schedules(pipeline_uuids):
+            a = aliased(PipelineSchedule, name='a')
+            result = (
+                PipelineSchedule.
+                select(*[
+                    a.created_at,
+                    a.id,
+                    a.name,
+                    a.pipeline_uuid,
+                    a.schedule_interval,
+                    a.schedule_type,
+                    a.start_time,
+                    a.status,
+                    a.updated_at,
+                ]).
+                filter(a.pipeline_uuid.in_(pipeline_uuids))
+            ).all()
+            return group_by(lambda x: x.pipeline_uuid, result)
+
+        mapping = {}
+        if include_schedules:
+            mapping = query_pipeline_schedules(pipeline_uuids)
+
+        for pipeline in pipelines:
+            schedules = []
+            if mapping.get(pipeline.uuid):
+                schedules = mapping[pipeline.uuid]
+            pipeline.schedules = schedules
+
+        return self.build_result_set(
+            pipelines,
+            user,
+            **kwargs,
+        )
+
+    @classmethod
+    def create(self, payload, user, **kwargs):
+        clone_pipeline_uuid = payload.get('clone_pipeline_uuid')
+        name = payload.get('name')
+        pipeline_type = payload.get('type')
+
+        if clone_pipeline_uuid is None:
+            pipeline = Pipeline.create(
+                name,
+                pipeline_type=pipeline_type,
+                repo_path=get_repo_path(),
+            )
+        else:
+            source = Pipeline.get(clone_pipeline_uuid)
+            pipeline = Pipeline.duplicate(source, name)
+
+        return self(pipeline, user, **kwargs)
+
     @classmethod
     async def member(self, pk, user, **kwargs):
         pipeline = await Pipeline.get_async(pk)
